feat: 마이페이지 내 프로필 화면 타임라인 컴포넌트 변경
