+++ src/assets/icons/MoveRightIcon.tsx
@@ -1,9 +1,13 @@
-export const MoveRightIcon = () => {
+type MoveRightIconProps = {
+  size?: number;
+};
+
+export const MoveRightIcon = ({ size = 24 }: MoveRightIconProps) => {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
+      width={size}
+      height={size}
       viewBox="0 0 24 24"
       fill="none"
     >
@@ -15,5 +19,5 @@ export const MoveRightIcon = () => {
         strokeLinejoin="round"
       />
     </svg>
-  )
-}
+  );
+};
